Implement simple logic for automatic ServingRuntime selection
An InferenceService is allowed to omit specifying a runtime. This should lead to automatic selection of a ServingRuntime from an allowed pool of ServingRuntimes.

These changes implement automatic selection of a ServingRuntime: if there are several runtimes available, order priority is by creation date (most recent first). If several runtimes have the same creation date, priority is by alphabetical order.