feat: Add creation of VirtualService if service-mesh enabled
Check the Model Mesh namespace to check if the namespace is
opendatahub.io/service-mesh=true. If so, create a VirtualService
attached to a Gateway when `enable-route` is set to true, else
create a VirtualService for intenral traffic.

This allows the Model Mesh Server to serve requests via the
Istio Mesh.

* restructured test suite to use unique namespaces pr test
* change builder image to publically accessible registry
* remove MeshEnabled static bool. Now determined dynamically