Add support for traffic splitting
Traffic splitting support depends on Service Mesh. When a Service Mesh is enabled, VirtualServices are created to route traffic to ModelMesh and OpenShift Routes are no longer created.

It is possible to do infer requests to specific InferenceServices. Annotations and Labels are used to group more than one InferenceService to create a logical group where traffic splitting would be done.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>