Fix reconcile loop
This fixes a reconcile loop. The `AuthConfig` from the template would end with `subresource.value` and `name.value` fields set to `null`, while the cluster stores these fields with empty strings. This would be detected as a diff at reconciliation and the `AuthConfig` would be re-applied causing a loop both in `odh-model-controller` and `authorino` controller.

Adding the empty fields to the template stops the loop.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>