Fix Authorization capability detection for manual setups
The authorization capability of ODH platform can either be configured by the ODH operator (managed) or manually configured by the user (unmanaged).

ODH operator was correctly detecting the availability of the authorization capability for managed setup, but it was not being correctly detected in the unmanaged setup.

This fixed odh-model-controller to detect the authorization capability for both managed and unmanaged configurations.

Fixes https://issues.redhat.com/browse/RHOAIENG-6430

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>